# driver-service/docs/swagger.yaml
basePath: /
definitions:
  dto.DriverCreateRequest:
    properties:
      carBrand:
        type: string
      carModel:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      lat:
        type: number
      lon:
        type: number
      plate:
        type: string
      taxiType:
        type: string
    type: object
  dto.DriverUpdateRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      location:
        $ref: '#/definitions/dto.Location'
      plate:
        type: string
    type: object
  dto.Location:
    properties:
      lat:
        type: number
      lon:
        type: number
    type: object
  dto.NearbyDriverResponse:
    properties:
      distanceKm:
        description: km cinsinden mesafe
        type: number
      firstName:
        type: string
      lastName:
        type: string
      plate:
        type: string
    type: object
  dto.RateDriverRequest:
    properties:
      score:
        type: number
    type: object
  dto.UpdateLocationRequest:
    properties:
      lat:
        type: number
      lon:
        type: number
    type: object
  dto.UpdateStatusRequest:
    properties:
      status:
        type: string
    type: object
  models.Driver:
    properties:
      carBrand:
        type: string
      carModel:
        type: string
      createdAt:
        type: string
      distance:
        type: number
      firstName:
        type: string
      id:
        description: Belgeyi mongo'ya "primitive _id" olarak kaydeder. - id(string)
          - objectID tipine çevirir.
        type: string
      lastName:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      plate:
        type: string
      rating:
        description: ortalama puan = 4.8 gibi. float döndüğünden dolayı.
        type: number
      ratingCount:
        description: kaç kişinin oy kullandığını tutar.
        type: integer
      status:
        description: Sürücünün o anki durumu. json:"status" -> API'nin yanıtını döner,
          bson:"status" -> veritabanında bu ad ile saklanır.
        type: string
      taxiType:
        type: string
      updatedAt:
        type: string
    type: object
  models.Location:
    properties:
      coordinates:
        description: '[longitude, latitude]'
        items:
          type: number
        type: array
      type:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    name: API Support
  description: Driver service içerisinde sürücü kayıt, arama, getirme, silme, durum
    güncelleme, status durumu, puanlamaya göre hesaplama yapan işlemleri yönetir.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Bitaksi Driver Service API
  version: "1.0"
paths:
  /driver/{id}:
    put:
      consumes:
      - application/json
      description: Driver'ın ad, soyad, araç ve plaka bilgilerini günceller.
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver güncelleme bilgileri
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DriverUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Hatalı İstek - StatusBadRequest
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sürücü Bulunamadı - StatusNotFound
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ID'ye göre Driver Günceller
      tags:
      - Drivers
  /drivers:
    get:
      consumes:
      - application/json
      description: Kayıtlı olan tüm driverları sayfalı yani (pagination) şeklinde
        listeler.
      parameters:
      - description: 'Page Number (Default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page Size (Default: 20)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'drivers: [], total: int, page: int'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Sunucu Hatası - Status Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Tüm Driverları Listele
      tags:
      - Drivers
    post:
      consumes:
      - application/json
      description: Sisteme yeni bir taksi sürücüsü kaydeder.
      parameters:
      - description: Sürücü Bilgileri
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DriverCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Hatalı İstek
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Plaka zaten kayıtlı
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Yeni Driver Oluştur
      tags:
      - Drivers
  /drivers/{id}:
    delete:
      consumes:
      - application/json
      description: API'den gelen isteğe göre ID ile eşleşen Driver'ı kalıcı olarak
        silme işlemi yapar.
      parameters:
      - description: Sürücü ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Geçersiz ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sürücü Bulunamadı
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ID'ye göre Driver Siler
      tags:
      - Drivers
    get:
      consumes:
      - application/json
      description: ID'si verilen driver'ın bilgilerini detaylı bir şekilde dönen metod.
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Geçersiz ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Sürücü bulunamadı
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ID'ye göre Driver Detayı Getir
      tags:
      - Drivers
  /drivers/{id}/location:
    put:
      consumes:
      - application/json
      description: Sürücünün hareket halinde konumunu güncelleyecek bir metod.
      parameters:
      - description: Sürücü ID
        in: path
        name: id
        required: true
        type: string
      - description: Konum Bilgisi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLocationRequest'
      responses:
        "200":
          description: Başarı Mesajı
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Hatalı İstek
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sürücü Bulunamadı
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Driver'dan Location Bilgisi Alma
      tags:
      - Driver
  /drivers/{id}/rate:
    post:
      consumes:
      - application/json
      description: Bir sürücüye 1-5 arası puan verir.
      parameters:
      - description: Sürücü ID
        in: path
        name: id
        required: true
        type: string
      - description: Puan Bilgisi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Sürücüyü Puanla
      tags:
      - Drivers
  /drivers/{id}/status:
    patch:
      consumes:
      - application/json
      description: Driver'ın uygulamayı başlatması, müşteri olması veya olmaması durumunu
        (available, busy or offline) günceller.
      parameters:
      - description: Sürücü ID
        in: path
        name: id
        required: true
        type: string
      - description: Durum Bilgisi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message ve status durumunu döner.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: geçersiz durum bilgisi
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Driver'ın Status Bilgilerini Günceller
      tags:
      - Drivers
  /drivers/nearby:
    get:
      consumes:
      - application/json
      description: Belirtilen konum ve yarıçaptaki Müsait (Available) taksileri listeler.
      parameters:
      - description: Enlem (Latitude)
        in: query
        name: lat
        required: true
        type: number
      - description: Boylam (Longitude)
        in: query
        name: lon
        required: true
        type: number
      - description: 'Yarıçap (km) - Default: 6'
        in: query
        name: radius
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.NearbyDriverResponse'
            type: array
      summary: Yakındaki Taksileri Getir
      tags:
      - Drivers
schemes:
- http
swagger: "2.0"